{kustomizations,environments}: initial commit
**digitalocean-cloud-controller-manager:**

This is the component that controls access to networking, including load
balancers that are required by TCP services and ingress-nginx, which
then manages load balancers for HTTP.

**digitalocean-csi:**

This is the component that manages storage for the cluster,
automatically creating volumes in DigitalOcean and automatically
attaching them to machines.

This component is not properly tested and I am not 100% sure that they
properly mount.

**ingress-nginx:**

This is the HTTP load balancer. Only one LoadBalancer Service is
required to use HTTP load balancing to multiple Pods over various
different domain names.

cert-manager and external-dns can be used to create certificates for
ingress-nginx and DNS records for the ingress-nginx LoadBalancer.